Manually implement event handlers, no heap alloc
Due to passing ReadOnlySpan, we needed to ToArray in order to pass it on the heap as an event argument.
Extract that behavior to an interface and replace the event handlers as a list. Checking the docs, appears they use a linkedlist, but this is sufficient for our use case.