Add some more utility commands to work with tables
ftwe ->For each entry in the array, finds a table at the specified field that is not empty
ftwes -> Find table with entries (display many form of ftwe)
ftowf -> For each entry in the array, get a table at the specified table field. For each object in the table find a value at the specified field that is not empty. (Useful when most tables only have a single object, but the root table has many)